+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# GEF Classic 3.24.0
+
+## Draw2D
+
+## GEF
+
+## Zest
+
 # GEF Classic 3.23.0
 
 ## Draw2D